types: Moved some Loader types from core to types
