+++ core/src/backend/navigator.rs
@@ -1,7 +1,7 @@
 //! Browser-related platform functions
 
-use crate::loader::Error;
 use indexmap::IndexMap;
+use ruffle_types::loader::Error;
 use ruffle_types::string::WStr;
 use std::future::Future;
 use std::path::{Path, PathBuf};
